+++ CheckPoint/common/views.py
@@ -1,6 +1,10 @@
 import requests
 import markdown2
-from django.shortcuts import render
+from django.contrib.auth.views import LoginView
+from django.shortcuts import render, redirect
+from django.contrib.auth.decorators import login_required, user_passes_test
+from django.contrib.auth.mixins import LoginRequiredMixin, UserPassesTestMixin
+from django.views.generic import TemplateView
 
 
 def home(request):
@@ -29,8 +33,14 @@ def website_news(request):
     return render(request, 'common/website-news.html', {'changelog_html': changelog_html})
 
 
-def admin_page(request):
-    return render(request, 'common/admin-panel.html')
+class AdminPanelView(LoginRequiredMixin, UserPassesTestMixin, TemplateView):
+    # if someone manually type admin-panel to forward him to login
+    template_name = 'common/admin-panel.html'
+    login_url = '/accounts/login/'
+
+    def test_func(self):
+        # check if user is staff
+        return self.request.user.is_staff
 
 
 def events(request):
@@ -83,4 +93,17 @@ def my_screenshots(request):
 
 def custom_404(request, exception=None):
     return render(request, 'common/404.html', status=404)
+
+
+class CustomAdminLoginView(LoginView):
+    template_name = 'accounts/login.html'
+
+    def dispatch(self, request, *args, **kwargs):
+        if request.user.is_authenticated:
+            if request.user.is_staff:
+                return redirect('/admin/')
+            else:
+                return redirect('profile')
+        return super().dispatch(request, *args, **kwargs)
+