+++ CheckPoint/roms/signals.py
@@ -1,0 +1,14 @@
+from django.db.models.signals import post_save, post_delete
+from django.dispatch import receiver
+from CheckPoint.roms.models import Review
+
+
+@receiver(post_save, sender=Review)
+def update_rom_rating_on_review_save(sender, instance, **kwargs):
+    instance.rom.update_rating()
+
+
+@receiver(post_delete, sender=Review)
+def update_rom_rating_on_review_delete(sender, instance, **kwargs):
+    instance.rom.update_rating()
+