+++ CheckPoint/urls.py
@@ -1,25 +1,11 @@
-"""
-URL configuration for CheckPoint project.
-
-The `urlpatterns` list routes URLs to views. For more information please see:
-    https://docs.djangoproject.com/en/6.0/topics/http/urls/
-Examples:
-Function views
-    1. Add an import:  from my_app import views
-    2. Add a URL to urlpatterns:  path('', views.home, name='home')
-Class-based views
-    1. Add an import:  from other_app.views import Home
-    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
-Including another URLconf
-    1. Import the include() function: from django.urls import include, path
-    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
-"""
 from django.contrib import admin
 from django.urls import path, include
 from django.conf import settings
 from django.conf.urls.static import static
+from CheckPoint.common.views import CustomAdminLoginView
 
 urlpatterns = [
+    path('admin/login/', CustomAdminLoginView.as_view(), name='admin_login'),
     path('admin/', admin.site.urls),
     path('', include('CheckPoint.common.urls'), name='home page'),
     path('accounts/', include('CheckPoint.accounts.urls'), name='accounts page'),